split mainjs to multiple files
